Made invariant checking explicit.
There are some invariants that I have deduced from looking at the game's data, such as "every object in
the game has an index property that looks a certain way". The functionality of this
project's code depends on the correctness of these assumptions. Their correctness is now checked against
the game data explicitly using the assertX() functions in quartersbrief.assert.js.