Changed getResearchPaths() to allow for upgrades to have a predecessor that has a different ucType.
This is not strictly speaking necessary, as it reflects a previous version of the game. At the moment, all ships in the game have only upgrades that have predecessors of the same ucType. However,

a) this allows for a future where Wargaming changes their mind about this again, and

b) much importantly, that every upgrade's predecessor has the same ucType is an assumption, and therefore would need to be validated by invariant checking. But because there are still legacy ships in the game data, such an invariant assertion would fail. Therefore, either some ships would need to be excluded from this particular invariant (seems unclean) or the invariant would have to not be checked at all (even worse).

Instead, I've just chosen to implement the old logic, since the new logic falls out of this as an edge case.
However, since this is an assumption, it