Added AccessorMixin.apply.
This is a more general method for manipulating deeply nested object properties, in that it allows to apply a passed function to any deeply nested property.
With this, get and set just become special cases where the function is either the identity function (that returns the value it was passed) and the
assignment function (that always returns a constant value regardless of arguments), resp.