Changed reference expansion from eager to lazy.
Before, all references were expanded eagerly by expandReferences(). This sometimes led to very large data objects that took a long time to retrieve and contained much data that was never actually going to be needed. Now, expandReferences changes the object's property when it detects a reference: The property is replaced with an accessor property that expands the reference on the first read and replaces the property with the result of the expansion.