GameObjectFactory._attachLabel now works recursively. Removed qb_ prefix from label property, and added invariant that checks that no game object has a property called "label".
The invariant is necessary because at the stage of game object creation when we attach the label, we are still working on the raw data. Moving label attachment to a later stage in the process - after conversion to `GameObject` - on the other hand would require _attachLabel to have knowledge of `GameObject` implementation details: how to read typeinfo.type. Attaching labels recursively to contained objects would also require directly reading through to `_data`.