Changed CDO back to direct writes on the data with backups.
The proxy-based version was proving far too slow, so this commit changes ComplexDataObject back to direct writing of the underlying data.

When multiply is called for the first time, it saves a backup of the original value. This information is used in .clear() and .freshCopy() to return the CDO to its original state, while at the same time a) avoiding the need to store the entire data again (which could be expected to roughly double the memory needs of the program) and b) prevents already expanded references to need to be expanded again.