Changed auto-update algorithm to work directly off the file system.
Previously, preferences.xml was read to determine current version. However, when the game client updates, this file does not get replaced - the new version only gets written with the next launch of the game.
The auto update algorithm now checks for the highest-numbered subfolder of World of Warships/bin/ and compares this to the last-used number.