Fixed a bug that caused get to sometimes return incorrect results with nested DataObjects.
The get call would be handed off to the nested DataObject with the whole remainder of the path instead of just the current key. This meant that if there were further elements to the path, get would try to traverse these, even though the hand-off had already completed their traversal. Now only a single step (currKey) is handed off to the nested data object.